requirements.txt
Flask
fpdf2
matplotlib